Update socket proxy to run via mix run --no-halt
It appears that running via a custom mix task meant it didn't pull in
the mix config for some reason. I couldn't get it to run with the
modified logger settings and new backend.
