+++ README.md
@@ -5,7 +5,7 @@ A simple app that listens on a port and proxies (and possibly fans out) data to 
 ## Usage
 
 ```
-$ mix socket_proxy --listen-port 8001 12.12.12.12:8002 13.13.13.13:8003
+$ env SOCKET_PROXY_LISTEN_PORT=8001 SOCKET_PROXY_DESTINATIONS="12.12.12.12:8002, 13.13.13.13:8003" mix run --no-halt
 ```
 
 ## Testing
@@ -31,7 +31,7 @@ $ mix fake_source --port 8000
 You can then run socket proxy to connect them all:
 
 ```
-$ mix socket_proxy --listen-port 8000 127.0.0.1:8080 127.0.0.1:8081
+$ env SOCKET_PROXY_LISTEN_PORT=8000 SOCKET_PROXY_DESTINATIONS="127.0.0.1:8080, 127.0.0.1:8081" mix run --no-halt
 ```
 
 Observe that data being generated by the two fake sources appear in both the fake servers. Try killing and restarting any combination of the windows and see that the data continues to flow.
